backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
redis==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9